requirements.txt
-i https://pypi.org/simple

# Core
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Admin==1.6.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-Swagger==0.2.14
SQLAlchemy>=2.0.16       
alembic==1.11.1
gunicorn==21.2.0
python-dotenv==1.0.0

# Utils
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.3.2
markupsafe==2.1.3
pyyaml==6.0.1
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
urllib3==2.2.1
Werkzeug==2.3.6
wtforms==3.1.2

# Optional
psycopg2-binary==2.9.9
cloudinary==1.39.0
certifi==2024.2.2
python-dotenv
flask-jwt-extended==4.6.0


